+++ package.json
@@ -4,12 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node dist/server.js",
-    "startDev": "node index.dev.js",
+    "start": "node index.js",
+    "startDev": "npm run start",
     "test": "VIRTUAL_HOST=127.0.0.1:1337 ./node_modules/.bin/mocha --compilers js:babel-core/register --require babel-polyfill $(find src -path '*spec*')",
-    "dev": "VIRTUAL_HOST=127.0.0.1:1337 npm run startDev & webpack-dev-server --progress --color",
+    "dev": "VIRTUAL_HOST=127.0.0.1:1337 npm start & webpack-dev-server --progress --color",
     "build": "npm run buildClient && npm run buildServer",
-    "buildClient": "./node_modules/.bin/webpack -p --define process.env.NODE_ENV='\"production\"' --progress --colors --config webpack.production.config.js",
+    "buildClient": "./node_modules/.bin/webpack -p --define process.env.NODE_ENV='\"production\"' --progress --config webpack.production.config.js",
     "buildServer": "./node_modules/.bin/webpack -p --define process.env.NODE_ENV='\"production\"' --progress --config webpack.server.config.js"
   },
   "keywords": [],
@@ -41,8 +41,9 @@
     "redux-mock-store": "^1.0.3",
     "sinon": "^1.17.4",
     "sinon-chai": "^2.8.0",
-    "webpack": "^2.1.0-beta.12",
-    "webpack-dev-server": "^2.1.0-beta.0",
+    "uglify-js": "github:mishoo/UglifyJS2#harmony",
+    "webpack": "^1.13.1",
+    "webpack-dev-server": "^1.14.1",
     "webpack-node-externals": "^1.2.0",
     "xo": "^0.15.1"
   },
