+++ src/actions/index.js
@@ -52,6 +52,8 @@ export function fetchEvent(event) {
 
     const eventInfo = await fetch(`${host}/api/events/${event}/info`)
       .then((response) => {
+        // console.log(response.status);
+        // console.log(response);
         if (response.status >= 400) {
           throw new Error('Bad response from server');
         }
@@ -60,6 +62,8 @@ export function fetchEvent(event) {
 
     const eventPosts = await fetch(`${host}/api/events/${event}/tweets`)
       .then((response) => {
+        // console.log(response.status);
+        // console.log(response);
         if (response.status >= 400) {
           throw new Error('Bad response from server');
         }
