+++ webpack.config.js
@@ -2,7 +2,6 @@ var path = require('path');
 var webpack = require('webpack');
 
 module.exports = {
-  // entry: './src/client.js',
   entry: [
     'webpack-dev-server/client?http://127.0.0.1:3001/',
     'webpack/hot/only-dev-server',
@@ -10,7 +9,6 @@ module.exports = {
     './src/client.js'
   ],
   output: {
-    // filename: './src/dist/bundle.js'
     path: path.join(__dirname, 'src/dist/'),
     filename: 'bundle.js'
   },
@@ -23,11 +21,7 @@ module.exports = {
       {
         test: /\.jsx?$/,
         exclude: /(node_modules|bower_components)/,
-        // loader: 'babel', // 'babel-loader' is also a legal name to reference
         loaders: ['react-hot', 'babel'] // ,
-        // query: {
-        //   presets: ['es2015', 'react']
-        // }
       }
     ]
   },
@@ -40,7 +34,7 @@ module.exports = {
   devServer: {
     hot: true,
     proxy: {
-      '*': 'http://127.0.0.1:' + (process.env.PORT) || 3001
+      '*': 'http://127.0.0.1:' + 3001
     },
     host: '127.0.0.1'
   }
